+++ lib/rabl.rb
@@ -2,6 +2,7 @@ require 'active_support'
 require 'active_support/core_ext/string/inflections'
 require 'active_support/core_ext/object/blank'
 require 'active_support/core_ext/hash/reverse_merge'
+require 'active_support/core_ext/hash/except'
 require 'active_support/core_ext/hash/slice'
 
 require 'rabl/version'