+++ test/builder_test.rb
@@ -23,21 +23,21 @@ context "Rabl::Builder" do
 
   context "#to_hash" do
     context "when given a simple object" do
-      setup { builder({ :attributes => { :name => :name } }) }
+      setup { builder({ :attributes => { :name => {} } }) }
       asserts "that the object is set properly" do
         topic.build(User.new, :root_name => "user")
       end.equivalent_to({ "user" => { :name => "rabl" } })
     end
 
     context "when given an object alias" do
-     setup { builder({ :attributes => { :name => :name } }) }
+     setup { builder({ :attributes => { :name => { :as => :foo } } }) }
       asserts "that the object is set properly" do
         topic.build(User.new, :root_name => "person")
-      end.equivalent_to({ "person" => { :name => "rabl" } })
+      end.equivalent_to({ "person" => { :foo => "rabl" } })
     end
 
     context "when specified with no root" do
-      setup { builder({ :attributes => { :name => :name } }) }
+      setup { builder({ :attributes => { :name => { :as => :name } } }) }
       asserts "that the object is set properly" do
         topic.build(User.new, :root => false)
       end.equivalent_to({ :name => "rabl" })
@@ -46,7 +46,7 @@ context "Rabl::Builder" do
 
   context "#attribute" do
     asserts "that the node" do
-      build_hash @user, :attributes => { :name => :name, :city => :city }
+      build_hash @user, :attributes => { :name => {}, :city => { :as => :city } }
     end.equivalent_to({:name => 'rabl', :city => 'irvine'})
 
     asserts "that with a non-existent attribute the node" do